refactor(inference): split the managed helpers out of models.rs
CI's "Enforce OpenHuman Rust file layout" step rejected the branch:

    src/openhuman/inference/provider/ops/models.rs: 777 lines (limit 750)

The three helpers added for the managed catalog pushed the file past the cap.
Moved them to `models_part_01.rs` and `include!`d it, matching the pattern
`http_error.rs` already uses in the same directory. 777 -> 688 lines; no
behaviour change.

`include!` is textual, so the helpers stay private to the `models` module and
the existing `#[path = "models_tests.rs"]` test module still sees them without
any visibility change. It also means only `//` comments are legal in the part
file — `//!` inner doc comments are rejected mid-file with E0753.
